Split the storage module into several submodules
