+++ src/chunks/storage/cow.rs
@@ -1,0 +1,40 @@
+use ndarray::{Array, ArrayView, Dimension, Ix1};
+
+/// Copy-on-write wrapper for `Array`/`ArrayView`.
+///
+/// The `CowArray` type stores an owned array or an array view. In
+/// both cases a view (`as_view`) or an owned array (`into_owned`) can
+/// be obtained. If the wrapped array is a view, retrieving an owned
+/// array will copy the underlying data.
+pub enum CowArray<'a, A, D> {
+    Borrowed(ArrayView<'a, A, D>),
+    Owned(Array<A, D>),
+}
+
+impl<'a, A, D> CowArray<'a, A, D>
+where
+    D: Dimension,
+{
+    pub fn as_view(&self) -> ArrayView<A, D> {
+        match self {
+            CowArray::Borrowed(borrow) => borrow.view(),
+            CowArray::Owned(owned) => owned.view(),
+        }
+    }
+}
+
+impl<'a, A, D> CowArray<'a, A, D>
+where
+    A: Clone,
+    D: Dimension,
+{
+    pub fn into_owned(self) -> Array<A, D> {
+        match self {
+            CowArray::Borrowed(borrow) => borrow.to_owned(),
+            CowArray::Owned(owned) => owned,
+        }
+    }
+}
+
+/// 1D copy-on-write array.
+pub type CowArray1<'a, A> = CowArray<'a, A, Ix1>;
